Cleanup, refactoring and optimizations

+++ ProjectEuler/Math/Sequences/ArithmeticSequence.cs
@@ -2,21 +2,21 @@
 {
     public class ArithmeticSequence
     {
-        private readonly decimal _commonDifference;
-        private readonly decimal _initialTerm;
+        private readonly long _commonDifference;
+        private readonly long _initialTerm;
 
-        public ArithmeticSequence(decimal initialTerm, decimal commonDifference)
+        public ArithmeticSequence(long initialTerm, long commonDifference)
         {
             _commonDifference = commonDifference;
             _initialTerm = initialTerm;
         }
 
-        public decimal GetTerm(int termIndex)
+        public long GetTerm(int termIndex)
         {
             return _initialTerm + _commonDifference * (termIndex - 1);
         }
 
-        public decimal GetSum(int termsCount)
+        public long GetSum(int termsCount)
         {
             return termsCount * (_initialTerm + GetTerm(termsCount)) / 2;
         }
